requirements.txt
langgraph==0.2.61
langchain-community==0.3.14
langchain-openai==0.2.14
tavily-python==0.5.0
langchain_chroma
openai==1.59.5
langchain==0.3.14
requests==2.32.3
pydantic==2.10.3
plotly==5.24.1
numpy
pandas
chromadb==0.5.23
pyngrok==7.2.3
st-clickable-images
unstructured[all-docs]==0.16.16
PyMuPDF==1.25.3
pyngrok==7.2.3
python-dotenv==1.0.1
streamlit==1.42.0
pytesseract==0.3.13